ai seems to work had to redo some things cause I forgot I would use polymorphsim for AI .... but working now still havent figured out how to initialise unique_ptr from initialiser list ... will have to research on this

+++ Player.cpp
@@ -119,7 +119,7 @@ int Ai_easy::play_turn(Board &board_to_play) {
                                                       getColor());
     board_to_play.set_color(board_to_play.getBase().first, board_to_play.getBase().second, getColor());
     board_to_play.change_color(flip_coordinates, getColor());
-    return flip_coordinates.size();
+    return (int) flip_coordinates.size();
 
 }
